Completed list display of ReservekraftBud with all hours of the day.
* Amended service with reading of bid price
* Added functionality to find volume for each hour
* Moved EdielService to client-package to be able to use in while testing client
* Using Availability instead of Duration for Varighet
* Test support: Insert test-data and read bid while starting UI
* Bug: Avoid duplicate reservekraftbud due to join with Volumperiode (Hibernate issue)
* Display contents of search panel
*

+++ larm-ediel-parser/src/main/java/no/statnett/larm/ediel/RngSegment.java
@@ -15,8 +15,8 @@ public class RngSegment extends EdifactSegment {
         return getElementComponent(1, 1);
     }
 
-    public Long getQuantity() {
-        return Long.valueOf(getMinimum());
+    public Integer getQuantity() {
+        return Integer.valueOf(getMinimum());
     }
 
     public void setQuantity(Long data) {
